gpu-service/requirements.txt
# WhisperX GPU transcription service
# Install with: pip install -r requirements.txt
#
# PyTorch — install BEFORE the rest, matching your CUDA version:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
#   (replace cu121 with your CUDA version — check with: nvidia-smi)
#
# Then: pip install -r requirements.txt

whisperx
pyannote.audio>=3.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart
python-dotenv
httpx
